Clean up splint and compilation warnings, fix minor gpsdecode option bugs.

+++ gpsd.c
@@ -1386,13 +1386,13 @@ static void handle_control(int sfd, char *buf)
 		/* NOTE: this destroys the original buffer contents */
 		st = gpsd_hexpack(eq, eq, len);
 		if (st <= 0) {
-		    gpsd_report(LOG_INF,"<= control(%d): invalid hex string (error %zd).\n", sfd, st);
+		    gpsd_report(LOG_INF,"<= control(%d): invalid hex string (error %d).\n", sfd, st);
 		    ignore_return(write(sfd, "ERROR\n", 6));
                 }
 		else
 		{
 		    gpsd_report(LOG_INF,"<= control(%d): writing %d bytes fromhex(%s) to %s\n", sfd, st, eq, stash);
-		    ignore_return(write(chp->gpsdata.gps_fd, eq, st));
+		    ignore_return(write(chp->gpsdata.gps_fd, eq, (size_t)st));
 		    ignore_return(write(sfd, "OK\n", 3));
 		}
 	    } else {
